refactor: reorganize service modules into subdirectories
