+++ packages/scaffold/src/index.ts
@@ -1,7 +1,7 @@
 export { ModuleCatalog } from "./catalog/ModuleCatalog";
 export { TargetCatalog } from "./catalog/TargetCatalog";
 export { ApplyService } from "./service/apply/ApplyService";
-export { BlueprintService } from "./service/BlueprintService";
-export { ContributionResolver } from "./service/ContributionResolver";
-export { PlanService } from "./service/PlanService";
+export { BlueprintService } from "./service/blueprint/BlueprintService";
+export { ContributionResolver } from "./service/plan/ContributionResolver";
+export { PlanService } from "./service/plan/PlanService";
 export { ScaffoldFormatter } from "./service/ScaffolFormatter";
