+++ packages/scaffold/src/service/plan/RepoSnapshotService.ts
